+++ CloudServer/src/Core/Encryption/EncryptionAPI.cpp
@@ -16,7 +16,7 @@ std::string EncryptionAPI::Encrypt(string data, string key)
 {
 	unsigned char iv[] = { 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff };
 	unsigned int out_len;
-	unsigned char* encrypted_data = aes.EncryptCBC((unsigned char*)data.c_str(), data.size(), (unsigned char*)key.c_str(), iv, out_len);
+	unsigned char* encrypted_data = aes.EncryptCBC((unsigned char*)data.c_str(), static_cast<unsigned int>(data.size()), (unsigned char*)key.c_str(), iv, out_len);
 
 	std::string encrypted = "";
 	encrypted.append((const char*)encrypted_data, static_cast<size_t>(out_len));
@@ -34,7 +34,7 @@ std::string EncryptionAPI::Decrypt(const string& data, string key)
 	}
 
 	unsigned char iv[] = { 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff };
-	unsigned char* decrypted_data = aes.DecryptCBC((unsigned char*)copy_data.c_str(), copy_data.size(), (unsigned char*)key.c_str(), iv);
+	unsigned char* decrypted_data = aes.DecryptCBC((unsigned char*)copy_data.c_str(), static_cast<unsigned int>(copy_data.size()), (unsigned char*)key.c_str(), iv);
 
 	std::string decrypted = "";
 	decrypted.append((const char*)decrypted_data, copy_data.size());
